Deprecate / update specific step across 8 features
https://github.com/aldefouw/rctf/blob/b87b94b7fe286682ff11efe8bce9ff9692d80243/step_definitions/interactions.js#L677-L704

This bulk change to 8 feature tests supports the ability to deprecate the specific step definition at the link above for several reasons:

1) This step definition contributes to Gherkin confusion by allowing "span" elements.

The span element is HTML jargon.  It's not something a REDCap SME will know unless they understand HTML, which defeats part of the purpose of Gherkin.

This was originally a work-around step definition and was never removed.

2) Removing this step adds clarity to the 8 features that currently use it

It's much clearer to say "button" instead of "span" to click on the "Choose action for record" element.  This change is reflected in this commit.

3) This step definition is redundant - other steps do virtually the same thing

There are other steps available that do virtually the same thing minus the "span" option.  This will reduce our code maintenance surface by 27 lines and 1 step definition.